+++ service/src/main/java/com/scrm/service/service/impl/CommunicationServiceImpl.java
@@ -50,7 +50,7 @@ public class CommunicationServiceImpl implements CommunicationService {
 
     @Override
     @Transactional//开启事务
-    public String deleteCommunication(Integer customerId) throws Exception{
+    public String deleteCommunication(Long customerId) throws Exception{
         int result = communicationDao.deleteCommunication(customerId);
         if (result < 1) {
             throw new Exception("删除失败");
@@ -59,7 +59,7 @@ public class CommunicationServiceImpl implements CommunicationService {
     }
 
     @Override
-    public Integer deleteCommunicationLog(Integer id) throws Exception {
+    public Integer deleteCommunicationLog(Long id) throws Exception {
         int result = communicationDao.deleteCommunicationLog(id);
         return null;
     }
